Update Godeps for libvirt
Fix missing commit ID 057f70a36a6b5
