+++ Godeps/_workspace/src/github.com/rgbkrk/libvirt-go/README.md
@@ -1,3 +1,7 @@
+# Carving out the future
+
+Go 1.6 is incompatible with libvirt's implementation of Domain Events (those things that callback to your code to let you know something changed in a domain's state), so builds of this under golang 1.6 will not include Domain Events unless someone comes up with a workaround.
+
 # libvirt-go
 
 [![Build Status](http://ci.serversaurus.com/github.com/alexzorin/libvirt-go/status.svg?branch=master)](http://ci.serversaurus.com/github.com/alexzorin/libvirt-go)
@@ -31,7 +35,7 @@ Previously there was support for libvirt 0.9.8 and below, however this is no lon
 
 ## Documentation
 
-* [api documentation for the bindings](http://godoc.org/github.com/alexzorin/libvirt-go)
+* [api documentation for the bindings](http://godoc.org/github.com/rgbkrk/libvirt-go)
 * [api documentation for libvirt](http://libvirt.org/html/libvirt-libvirt.html)
 
 ## Contributing
