+++ Godeps/_workspace/src/github.com/rgbkrk/libvirt-go/cfuncs.go
@@ -1,14 +1,18 @@
+//+build !go1.6
+
 package libvirt
+
+/*
+ * Golang 1.6 doesn't support C pointers to go memory.
+ * A hacky-solution might be some multi-threaded approach to support domain events, but let's make it work
+ * without domain events for now.
+ */
 
 /*
 #cgo LDFLAGS: -lvirt 
 #include <libvirt/libvirt.h>
 #include <libvirt/virterror.h>
 #include <stdlib.h>
-
-void virErrorFuncDummy(void *userData, virErrorPtr error)
-{
-}
 
 int domainEventLifecycleCallback_cgo(virConnectPtr c, virDomainPtr d,
                                      int event, int detail, void *data)
