+++ Godeps/_workspace/src/github.com/rgbkrk/libvirt-go/secret.go
