+++ Godeps/_workspace/src/github.com/rgbkrk/libvirt-go/storage_pool.go
