+++ Godeps/_workspace/src/github.com/rgbkrk/libvirt-go/vagrant/master/Vagrantfile
