+++ Godeps/_workspace/src/github.com/rgbkrk/libvirt-go/vagrant/vagrant-provision-reboot-plugin.rb
